requirements.txt
modal==0.57.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
scipy==1.11.3
trimesh==3.23.5
pytorch3d==0.7.5
face-alignment==1.3.5
insightface==0.7.3
onnxruntime-gpu==1.16.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
pydantic==2.4.2
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
flame-pytorch @ git+https://github.com/soubhiksanyal/FLAME_PyTorch.git